% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialML.R
\name{assignTrainAndTestSet}
\alias{assignTrainAndTestSet}
\title{Divide the training and test set based on day number}
\usage{
assignTrainAndTestSet(directoryPath)
}
\description{
Divide the training and test set based on day number
}

