% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainValidTestSetCreate.R
\name{coupleImagesAndMeteoToDate}
\alias{coupleImagesAndMeteoToDate}
\title{Couple cameras and KNMI nearby stations}
\usage{
coupleImagesAndMeteoToDate(dateStrInitial, dateStrFinal, dbConfigDir, maxDist)
}
\arguments{
\item{dbConfigDir}{String of path with directory containing the DB param access config file}

\item{maxDist}{Numerical containing max distance from KNMI station}

\item{dateStr}{String of latest date do use for fetching data}
}
\description{
Couple cameras and KNMI nearby stations
}
