% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainValidTestSetCreate.R
\name{createTrainValidTestSetsBinary}
\alias{createTrainValidTestSetsBinary}
\title{Create training, validation and test datasets}
\usage{
createTrainValidTestSetsBinary(dataDir, dateMax = "'2018-02-28 00:00:00'",
  dbConfigDir)
}
\arguments{
\item{dataDir}{String of directory containing the cameras}

\item{dateMax}{String of latest date do use for fetching data}
}
\description{
Create training, validation and test datasets
}

