% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFeatureCompute.R
\name{featureExtraction}
\alias{featureExtraction}
\title{Saves the features of the images}
\usage{
featureExtraction(propertiesLocationsVect)
}
\arguments{
\item{propertiesLocationsVect}{vector containing the configuration/properties
of the location of the images}
}
\description{
Saves the features of the images
}

