% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainValidTestSetCreate.R
\name{imagesAndMeteoFogBinary}
\alias{imagesAndMeteoFogBinary}
\title{Assign images a binary fog property (foggy=TRUE/FALSE)}
\usage{
imagesAndMeteoFogBinary(dtImages, dtMeteo)
}
\arguments{
\item{dtImages}{Data table of images from the DB}

\item{dtMeteo}{Data table of meteo (visibility) from the DB}
}
\description{
Assign images a binary fog property (foggy=TRUE/FALSE)
}

