% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMeteoData.R
\name{ReadHumidityData}
\alias{ReadHumidityData}
\title{Read relative humidity data files}
\usage{
ReadHumidityData(filename)
}
\arguments{
\item{filenames}{List of filenames}
}
\value{
data.table
}
\description{
Read relative humidity data files
}

