% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMeteoData.R
\name{ReadWindData}
\alias{ReadWindData}
\title{Read wind data files}
\usage{
ReadWindData(filenames)
}
\arguments{
\item{filenames}{List of filenames}
}
\value{
data.table
}
\description{
Read wind data files
}

