% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFeatureCompute.R
\name{ReturnFeatures}
\alias{ReturnFeatures}
\title{Computes the features of the images}
\usage{
ReturnFeatures(filePath)
}
\arguments{
\item{filePath}{String of the file location}
}
\description{
Computes the features of the images
}

