% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationsCamerasCoupling.R
\name{coupleCamerasAndKNMInearStations}
\alias{coupleCamerasAndKNMInearStations}
\title{Couple cameras and KNMI nearby stations}
\usage{
coupleCamerasAndKNMInearStations(maxDistance = 5000, dbConfigDir)
}
\arguments{
\item{maxDistance}{Numeric distance radius from the KNMI station}

\item{dbConfigDir}{String of path with directory containing the DB param access config file}
}
\description{
Couple cameras and KNMI nearby stations
}

