% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainValidTestSetCreate.R
\name{coupleImagesAndMeteoToDate}
\alias{coupleImagesAndMeteoToDate}
\title{Couple cameras and KNMI nearby stations}
\usage{
coupleImagesAndMeteoToDate(dateStr, dbConfigDir)
}
\arguments{
\item{dateStr}{String of latest date do use for fetching data}

\item{dbConfigDir}{String of path with directory containing the DB param access config file}
}
\description{
Couple cameras and KNMI nearby stations
}

