% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFeatureCompute.R
\name{featureExtraction}
\alias{featureExtraction}
\title{Saves the features of the images}
\usage{
featureExtraction(propertiesLocationsVect)
}
\arguments{
\item{propertiesLocations}{data table containing the configuration/properties of that image}
}
\description{
Saves the features of the images
}

